+++ Core/Objects/Encounter.hpp
