+++ Core/Objects/EncounterSlot.cpp
