+++ Core/Objects/EncounterSlot.hpp
