+++ Core/Objects/LCRNG.cpp
