+++ Core/Objects/LCRNG.hpp
