+++ Core/Objects/MTRNG.cpp
