+++ Core/Objects/NatureLock.cpp
