+++ Core/Objects/NatureLock.hpp
