+++ Core/Objects/SFMT.cpp
