+++ Core/Objects/Utilities.cpp
