+++ Core/Objects/Utilities.hpp
