+++ main.cpp
@@ -1,12 +1,5 @@
-#include "mainwindow.h"
+#include "Forms/mainwindow.h"
 #include <QApplication>
-#include "Objects/NatureLock.hpp"
-#include "Objects/Utilities.hpp"
-#include "Objects/LCRNG.hpp"
-#include "Objects/SFMT.hpp"
-#include "Objects/TinyMT.hpp"
-#include <iostream>
-#include <boost/date_time/gregorian/gregorian_types.hpp>
 
 int main(int argc, char *argv[])
 {
